leetcode 两数之和

+++ main.c
@@ -5,7 +5,7 @@ int main()
 {
 	int a = -1;
 	unsigned int b = 1000;
-	int c = a < (long)b ? 1 : 0;
+	int c = a < b ? 1 : 0;
 	printf("hello git%d\n", c);
 	return 0;
 }
